Fix a race on stop/exit of a thread
`std::condition_variable::notify_one` was called not from under the
lock, this way it was impossible to guarantee that
`std::condition_variable` stays at `wait` in the `runLoop`. In this case
if `notify_one` is called between checking `getIsRunning()` and `wait`
(quite rare case, but happens now and then), than call to `notify_one`
will be missed.